+++ src/common/i18n/en.js
@@ -27,9 +27,9 @@ Validity.i18n = {
                 start_line: 'The starting date of the validity period (from infinite if nul)'
             },
             panel: {
-                confirm_mergeleft: 'Merging with previous period means keeping these current data, consolidating the validity periods. Are you sure ?',
-                confirm_mergeright: 'Merging with next period means keeping these current data, consolidating the validity periods. Are you sure ?',
-                confirm_remove: 'You are about to remove the validity period. Are you sure ?',
+                confirm_mergeleft: 'Merging with previous period means keeping these current data, consolidating the validity periods.<br />Are you sure ?',
+                confirm_mergeright: 'Merging with next period means keeping these current data, consolidating the validity periods.<br />Are you sure ?',
+                confirm_remove: 'You are about to remove the validity period.<br />Are you sure ?',
                 end_legend: 'Valid until',
                 from: 'From %s',
                 fromto: 'From %s to %s',